refactor: 각 턴의 상태가 추상클래스를 상속받도록 수정
